Rewrite .gitignore to actually ignore everything
`*` ignores everything in the root directory. `*/*` ignores everything
in every subdirectory. This also works for subdirectories with their own
.gitignores, such as an .idea directory.

Whitelist a file in the root directory:

    !file

Whitelist a file in a subdirectory:

    !dir
    !dir/file

Whitelist everything in a directory:

    !dir
    !dir/*

Recursively whitelist a directory:

    !dir
    !dir/**

You can't just do `!dir/file`, because `dir` needs to be whitelisted
separately.