Add demo Rocket backend + React frontend
The top-level directory is the Rust project (server).

The app/ directory is the React project (frontend).

The React app is the default created via `npx create-react-app app`.

To build this you'll need to install npm and rustup:

https://www.npmjs.com/get-npm

https://rustup.rs

Then switch to the nightly toolchain:

    rustup default nightly

Then you can build and run everything:

    cd app
    npm install
    npm run build
    cd ..
    cargo run

This will serve the app locally and you should be able to follow the
link that Rocket printed to your terminal. It will probably be something
like localhost:8000