Add database and POST request for signup
You'll need the `sqlite3` command installed on your development machine.
It might already be installed.

On launch, we create the file `database.db`. If the file already exists,
we use the existing one.

We then add a `User` table where each entry has a `username` and
`password_hash`. If the table already exists, we use the existing one.

To sign up, the client sends a POST request to /signup with JSON data in
the following form:

    {
        "username": "bob",
        "password_hash": 123
    }

This will insert a User with those exact fields.

Keep in mind that this is just a first version. This isn't necessarily
secure and we also don't check if the user already exists (it will just
insert a new user with the same name).