Add POST request for login
/login accepts a POST request with the same data as /signup:

    {
        username: "bob",
        password_hash: 123
    }

If the username and password match a row in the database, the username
is added to the cookies, so that in subsequent requests the server can
simply check the cookie to see who's logged in.

Notice that I've renamed the Signup type to Login - this is simply a
generalisation, since login and signup both take a username and
password. Login here is a noun.