Set default avatar and add /set_avatar request
Now User has an avatar column. It is never null because we supply a
default avatar. The default avatar is stored as static data via
include_bytes, but for some reason we have to convert it to a Vec. It
might be more efficient to convert it to a Vec at startup and just
reference that one Vec.

/set_avatar takes a POST request with Content-Type 'image/png'. If
nobody is logged in, the server will return "NotLoggedIn". The server
does not validate the PNG data. It's just some binary data which the
frontend will deal with.