Rewrite parallel test system
The old code actually had a data race between dir.exists() and
create_dir(). It was possible for two threads to do this at the same
time, meaning one thread would try to create the directory when it
already exists.

This new system is much simpler:
1. The test function calls setup!() which returns the client and test
   database.
2. setup_test_rocket takes a number for the database file. setup!()
   gives it line!(), which is the line number of whoever called
   setup!().
3. setup_test_rocket will remove the test directory and create a new
   one. This is done in a SETUP which is a Once (thread-safe function
   which is only ever called once). Then a new database file is created
   using the line number, making it unique.

The old system didn't even allow you to inspect the databases of tests,
because there was no way to find out which test created which file! With
this, the database filename corresponds to the line number in test.rs.