Implement voice chat for channels
- Multiple peers
- Join and leave freely

One issue with this is that if leaveVoice() is not called (e.g. due to
unplugging at the wall) the server won't know that the peer has left the
voice chat, which is a memory leak among other problems.

Also, the server totally trusts whatever clients send, which is in turn
sent to the other clients. It's therefore very easy to send bad data and
ruin the voice chat (and the internal state of the server).